Extract global setup logic to widget
